Fix cycle detection reset
Some tests were failing because cycles were being detected that
weren't actually cycles. The clearing logic for references needs
to actually walk the cycle to determine if it needs to be walked
further.

Previously, I thought it would be ok to short circuit, but because
we already short circuit if we detect states that haven't been marked,
we have to walk all the states that HAVE been marked, which may have
happened through a ref.